+++ pybop/__init__.py
@@ -180,6 +180,11 @@ from .plotting.plot_problem import quick_plot
 from .plotting.nyquist import nyquist
 
 #
+# Experimental
+#
+from .experimental.jax_costs import JaxSumSquaredError, JaxLogNormalLikelihood
+
+#
 # Remove any imported modules, so we don't expose them as part of pybop
 #
 del sys